Fix #4 ExtensionDataService
If agents have the [Allow scripts to access OAuth
token](https://docs.microsoft.com/en-us/vsts/build-release/concepts/process/phases#agent-phase)
option checked on all agents with select/swap tasks, then selection data
will be stored in the ExtensionDataService instead.  This should allow
different agents, running as the same user, to access the data.

+++ BlueGreenSwap/Swap-F5BlueGreen.ps1
@@ -11,20 +11,37 @@ begin {
     Import-Module $env:CURRENT_TASK_ROOTDIR\F5-LTM\F5-LTM.psm1 -Force
 }
 process {
-    $f5Selections = Get-ChildItem -Path $env:SYSTEM_WORKFOLDER -Filter ('F5BlueGreen-{0}-{1}-*.xml' -f $env:Release_ReleaseId,$env:Release_EnvironmentId)
+    $storagekey = 'F5BlueGreen-{0}-{1}' -f $env:Release_ReleaseId,$env:Release_EnvironmentId
+    if ($env:SYSTEM_TOKEN) {
+        # Base64-encodes the Personal Access Token (PAT)
+        $Auth = @{ Authorization = 'Basic {0}' -f $([System.Convert]::ToBase64String([System.Text.Encoding]::UTF8.GetBytes(":$($Env:SYSTEM_TOKEN)"))) }
+        $uri = '{0}/_apis/ExtensionManagement/InstalledExtensions/vercellj/f5-tasks/Data/Scopes/User/Me/Collections/%24settings/Documents?api-version=3.1-preview.1' -f ($Env:SYSTEM_TEAMFOUNDATIONSERVERURI -replace '^[^.]*\.','$0extmgmt.')
+        # Store Credentials in JSON as plain text for now.
+        $f5Selections = Invoke-RestMethod -Uri $uri -Method Get -Headers $Auth | Select-Object -ExpandProperty Value | Where-Object { $_.id -match "^$storagekey" }
+    } else {
+        $f5Selections = Get-ChildItem -Path $env:SYSTEM_WORKFOLDER -Filter ('{0}-*.xml' -f $storagekey)
+    }
     if ($f5Selections) {
         $swapcount = 0
         $f5Selections | ForEach-Object {
             try {
-                $BlueGreenData = Import-Clixml -Path $_.FullName
+                if ($env:SYSTEM_TOKEN) {
+                    $BlueGreenData = Invoke-NullCoalescing {$_.value} {$_}
+                    # Re-constitute credentials from plaintext
+                    $SecurePassword = ConvertTo-SecureString $BlueGreenData.Credentials.Password -AsPlainText -Force
+                    $BlueGreenData.Credentials = New-Object System.Management.Automation.PSCredential ($BlueGreenData.Credentials.UserName, $SecurePassword)
+                } else {
+                    $BlueGreenData = Import-Clixml -Path $_.FullName
+                }
                 $session = New-F5Session -LTMName $BlueGreenData.LTMName -LTMCredentials $BlueGreenData.Credentials -PassThrough
 
                 $BlueGreenData.VirtualServer.pool = $BlueGreenData.IdlePool.name
-                
-                # Remove persist property if it exists.  Inclusion causes Set-VirtualServer not to update without any indication.
-                if ($BlueGreenData.VirtualServer.persist) { $BlueGreenData.VirtualServer.PSObject.properties.remove('persist'); }
 
-                $UpdatedServer = $BlueGreenData.VirtualServer | Set-VirtualServer -F5Session $session -Name $BlueGreenData.VirtualServer.name -Application $BlueGreenData.Application -Partition $BlueGreenData.Partition -PassThru
+                ## Remove persist property if it exists.  Inclusion causes Set-VirtualServer not to update without any indication.
+                #if ($BlueGreenData.VirtualServer.persist) { $BlueGreenData.VirtualServer.PSObject.properties.remove('persist'); }
+
+                #$UpdatedServer = $BlueGreenData.VirtualServer | Set-VirtualServer -F5Session $session -Name $BlueGreenData.VirtualServer.name -Application $BlueGreenData.Application -Partition $BlueGreenData.Partition -PassThru
+                $UpdatedServer = Set-VirtualServer -F5Session $session -Name $BlueGreenData.VirtualServer.name -Application $BlueGreenData.Application -Partition $BlueGreenData.Partition -DefaultPool $BlueGreenData.IdlePool.name -PassThru
 
                 # Validate swap
                 if ($UpdatedServer.pool -ne $BlueGreenData.IdlePool.fullPath) {
@@ -50,7 +67,12 @@ process {
                         Write-Host ("##vso[task.logissue type=warning;] Failed to sync device to group {0}" -f $BlueGreenData.DeviceGroup)
                     }
                 }
-                Remove-Item -Path $_.FullName -Force
+                if ($env:SYSTEM_TOKEN) {
+                    $uri = '{0}/_apis/ExtensionManagement/InstalledExtensions/vercellj/f5-tasks/Data/Scopes/User/Me/Collections/%24settings/Documents/{1}?api-version=3.1-preview.1' -f ($Env:SYSTEM_TEAMFOUNDATIONSERVERURI -replace '^[^.]*\.','$0extmgmt.'),$_.id
+                    Invoke-RestMethod -Uri $uri -Method Delete -Headers $Auth
+                } else {
+                    Remove-Item -Path $_.FullName -Force
+                }
             } catch {
                 Write-Host ("##vso[task.logissue type=error;] Failed to swap the active pool for '{0}'; {1}" -f $BlueGreenData.VirtualServer.name,$_.Exception.Message)
                 throw New-Object -TypeName System.Exception -ArgumentList ("Failed to swap the active pool for '{0}'; {1}" -f $BlueGreenData.VirtualServer.name,$_.Exception.Message)
